src/Backend/backend/requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.1

# Jupyter and kernel management
jupyter-client==8.5.0
ipykernel==6.26.0
jupyter-core==5.5.0

# AI and OpenAI integration
openai==1.3.0
python-dotenv==1.0.0

# File handling and validation
python-multipart==0.0.6
aiofiles==23.2.1
python-magic==0.4.27  # For MIME type detection

# Logging and monitoring
python-json-logger==2.0.7
structlog==23.2.0

# Background tasks and scheduling
apscheduler==3.10.4
celery==5.3.4  # Optional for distributed task processing

# Security and authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# HTTP client and utilities
httpx==0.25.2
requests==2.31.0

# Data validation and serialization
marshmallow==3.20.1
validators==0.22.0

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# Async and concurrency
asyncio-mqtt==0.15.0  # Optional for MQTT support
aioredis==2.0.1  # Optional for Redis caching

# Development and testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2  # For testing async clients
faker==20.1.0  # For generating test data

# Code quality and formatting
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# Documentation
sphinx==7.2.6
sphinx-autodoc-typehints==1.25.2

# Environment and configuration
python-decouple==3.8
click==8.1.7  # For CLI commands

# Performance and monitoring
prometheus-client==0.19.0
psutil==5.9.6

# File format support
openpyxl==3.1.2  # Excel files
pandas==2.1.3  # Data manipulation
pillow==10.1.0  # Image processing
matplotlib==3.8.2  # Plotting
seaborn==0.13.0  # Statistical plotting

# Networking and protocols
websockets==12.0  # WebSocket support
sse-starlette==1.8.2  # Server-sent events

# Caching
redis==5.0.1
python-memcached==1.62

# Serialization
orjson==3.9.10  # Fast JSON serialization
msgpack==1.0.7

# Utility libraries
more-itertools==10.1.0
toolz==0.12.0
attrs==23.1.0

# OS-specific dependencies (commented out, install as needed)
# Windows
# pywin32==306  # Windows-specific utilities

# Linux/Unix
# python-prctl==1.8.1  # Process control on Linux

# macOS
# pyobjc-framework-Cocoa==10.0  # macOS-specific utilities

# Optional dependencies for advanced features
# Machine learning and data science
# scikit-learn==1.3.2
# numpy==1.25.2
# scipy==1.11.4
# tensorflow==2.14.0  # If ML features are needed
# torch==2.1.1  # Alternative ML framework

# Database drivers (choose based on your database)
# PostgreSQL
# psycopg2-binary==2.9.9
# asyncpg==0.29.0

# MySQL
# pymysql==1.1.0
# aiomysql==0.2.0

# MongoDB
# pymongo==4.6.0
# motor==3.3.2  # Async MongoDB driver

# Message queues and brokers
# pika==1.3.2  # RabbitMQ
# kafka-python==2.0.2  # Apache Kafka

# Distributed computing
# dask==2023.11.0  # Distributed computing
# ray==2.8.0  # Distributed AI/ML

# Additional FastAPI extensions
starlette-prometheus==0.9.0
fastapi-cache2==0.2.1
fastapi-limiter==0.1.5

# Email and notifications
# sendgrid==6.10.0
# twilio==8.11.0

# Cloud storage
# boto3==1.33.13  # AWS S3
# google-cloud-storage==2.10.0  # Google Cloud Storage
# azure-storage-blob==12.19.0  # Azure Blob Storage

# Configuration management
dynaconf==3.2.4

# Process management
supervisor==4.2.5

# Development tools
pre-commit==3.6.0
bandit==1.7.5  # Security linting 